+++ .ls-lint.yml
@@ -84,3 +84,6 @@ ls:
   libraries/api-client/src/userGroups:
     .dir: camelCase
     .ts: camelCase
+  libraries/api-client/src/self:
+    .dir: camelCase
+    .ts: camelCase
