+++ libraries/api-client/src/self/changePassword.ts
