+++ libraries/api-client/src/self/consent.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2021 Wire Swiss GmbH
+ * Copyright (C) 2018 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,8 +17,10 @@
  *
  */
 
-import {Consent} from './Consent';
+import {ConsentType} from './consentType';
 
-export interface ConsentResults {
-  results: Consent[];
+export interface Consent {
+  source: string;
+  type: ConsentType;
+  value: number;
 }
