+++ libraries/api-client/src/self/consentResults.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2021 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,15 +17,8 @@
  *
  */
 
-import {SSOSignature} from '../self/';
-import {ManagedSource, User} from '../user/';
+import {Consent} from './consent';
 
-export interface Self extends User {
-  locale: string;
-  /**
-   * What is the source of truth for this user; if it's SCIM
-   * then the profile can't be edited via normal means.
-   */
-  managed_by?: ManagedSource;
-  sso_id?: SSOSignature;
+export interface ConsentResults {
+  results: Consent[];
 }
