+++ libraries/api-client/src/self/consentType.ts
