+++ libraries/api-client/src/self/delete.ts
