+++ libraries/api-client/src/self/picture.ts
