+++ libraries/api-client/src/self/self.ts
@@ -17,10 +17,16 @@
  *
  */
 
-import {ConsentType} from './ConsentType';
+import {SSOSignature} from './ssoSignature';
 
-export interface Consent {
-  source: string;
-  type: ConsentType;
-  value: number;
+import {ManagedSource, User} from '../user/';
+
+export interface Self extends User {
+  locale: string;
+  /**
+   * What is the source of truth for this user; if it's SCIM
+   * then the profile can't be edited via normal means.
+   */
+  managed_by?: ManagedSource;
+  sso_id?: SSOSignature;
 }
