+++ libraries/api-client/src/self/ssoSignature.ts
